refactor: Move Footer
